examples/client-workflow/requirements.txt
bnbagent
httpx>=0.25.0
python-dotenv>=1.0.0
